refactor : comment entity change to board entity embedded collection

+++ src/backend/community-service/src/main/java/harmony/communityservice/board/comment/dto/SearchCommentResponse.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 
 @Builder(toBuilder = true)
 public record SearchCommentResponse(
-        Long commentId,
+        int commentId,
         Long boardId,
         String writerName,
         String comment,
